Major upgrade. New templater, speed increase x3, custom FormatCodes
Pod::To::HTML in this distribution was 3x slower than the legacy Pod::To::HTML. Now it is
just under 1% faster, whilst achieving more.

The speed increase is due to the new templater, which runs everything as raku subroutines.

The number of minimum templates was slightly reduced, and the possibility of custom FormatCodes has been included.